# server/spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Prompt Mint API
  version: "1.0.0"
  description: |
    Prompt Mint marketplace API — Express backend serving prompt listing,
    user management, governance, appeals, licensing, webhooks, versioning,
    notifications, analytics, and API key management.
  contact:
    name: Prompt Mint Team
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.promptmint.io
    description: Production server
paths:
...
  /api/webhooks:
    post:
      operationId: registerWebhook
      summary: Register a new webhook subscription
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                secret:
                  type: string
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
    get:
      operationId: getWebhook
      summary: Get webhook subscription details
      tags: [Webhooks]
      responses:
        "200":
          description: Webhook subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteWebhook
      summary: Delete a webhook subscription
      tags: [Webhooks]
      responses:
        "200":
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /api/webhooks/dead-letters/{id}/replay:
    post:
      operationId: replayWebhookDeadLetter
      summary: Replay a dead-letter webhook event
      description: >-
        Admin-token gated. Re-sends the stored envelope verbatim by default.
        Set refreshTimestamp to re-stamp the envelope timestamp with the
        current time before signing, which is required to redeliver an event
        that has aged past a receiver's acceptance window. deliveryId is
        preserved in both modes so a receiver that already processed the event
        can dedupe it.
      tags: [Webhooks]
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshTimestamp:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Replay attempted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  statusCode:
                    type: integer
                  error:
                    type: string
                  replayedAt:
                    type: string
                    format: date-time
                  refreshTimestamp:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/webhooks/replay/queue:
    get:
      operationId: getWebhookReplayQueue
      summary: Get the replay queue for a wallet's webhook
      description: >-
        Dead-lettered events for the wallet's subscription, each with an
        assessment of whether replaying it can succeed.
      tags: [Webhooks]
      parameters:
        - name: walletAddress
          in: query
          required: true
          schema:
            type: string
        - name: resolved
          in: query
          required: false
          schema:
            type: string
            enum: ["true", "false"]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Replay queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookReplayQueue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/webhooks/replay/preview:
    post:
      operationId: previewWebhookReplay
      summary: Preview the envelope for a webhook event
      description: >-
        Builds the exact body and header set a receiver would see for an event
        without delivering anything. Carries a fresh deliveryId and timestamp;
        it is not what a dead-letter replay re-sends.
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event]
              properties:
                event:
                  type: string
                data:
                  type: object
      responses:
        "200":
          description: Envelope preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookReplayPreview"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ─── Notifications ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: Admin API token from ADMIN_API_TOKEN env var
    apiKeyQuery:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key for programmatic access
  responses:
    BadRequest:
      description: Bad request — missing or invalid fields
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
...
  schemas:
...
    WebhookSubscription:
      type: object
      properties:
        _id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    WebhookPayload:
      type: object
      properties:
        version:
          type: integer
        schemaVersion:
          type: string
        event:
          type: string
        deliveryId:
          type: string
        timestamp:
          type: string
          format: date-time
        data:
          type: object
...
    WebhookReplayAssessment:
      type: object
      description: >-
        Per-row verdict on whether replaying a dead letter can succeed. The
        timestamp of a stored envelope never changes on a verbatim replay, so a
        dead letter older than acceptanceWindowSeconds is flagged `stale`:
        a receiver following server/docs/webhook-signatures.md is expected to
        reject it unless the replay is re-stamped.
      properties:
        deliveryId:
          type: string
        event:
          type: string
        originalTimestamp:
          type: string
        ageSeconds:
          type: integer
          nullable: true
        fingerprint:
          type: string
          description: SHA-256 of the canonical payload JSON; unchanged by a verbatim replay.
        replayable:
          type: boolean
        stale:
          type: boolean
        warnings:
          type: array
          items:
            type: string
            enum: [already_resolved, stale_event, unparseable_timestamp, subscription_missing, subscription_inactive]
    WebhookReplayQueueItem:
      type: object
      properties:
        id:
          type: string
        event:
          type: string
        attempts:
          type: integer
        lastError:
          type: string
          nullable: true
        lastStatusCode:
          type: integer
          nullable: true
        resolved:
          type: boolean
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        replayCount:
          type: integer
        lastReplayedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        payload:
          $ref: "#/components/schemas/WebhookPayload"
        subscription:
          type: object
          nullable: true
          description: Delivery target for this dead letter. Never contains the subscription secret.
          properties:
            id:
              type: string
            url:
              type: string
            active:
              type: boolean
        replay:
          $ref: "#/components/schemas/WebhookReplayAssessment"
    WebhookReplayQueue:
      type: object
      properties:
        summary:
          type: object
          properties:
            total:
              type: integer
            pending:
              type: integer
            resolved:
              type: integer
            replayable:
              type: integer
            stale:
              type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/WebhookReplayQueueItem"
        acceptanceWindowSeconds:
          type: integer
    WebhookReplayPreview:
      type: object
      description: >-
        The exact envelope and header set a receiver would see for an event,
        built without contacting anything. The preview always carries a fresh
        deliveryId and timestamp; replaying a dead letter re-sends the stored
        envelope instead.
      properties:
        event:
          type: string
        schemaVersion:
          type: string
        payload:
          $ref: "#/components/schemas/WebhookPayload"
        body:
          type: string
          description: Exact request body a receiver would verify.
        headers:
          type: object
          additionalProperties:
            type: string
        acceptanceWindowSeconds:
          type: integer

    # ── Notifications ─────────────────────────────────────────────────────────
...